Migrate um/vswriter to 0.3
