+++ Cargo.toml
@@ -185,6 +185,7 @@ usbspec = []
 usp10 = []
 vss = []
 vsserror = []
+vswriter = []
 werapi = []
 winbase = []
 wincodec = []
